fixed bug in geo indexed raster batch processing where file id was truncated to 32 bits

+++ packages/geo/package/GeoIndexedRaster.h
@@ -152,7 +152,7 @@ class GeoIndexedRaster: public RasterObject
         } sample_collector_t;
 
         /* Map of raster file id to list of points to be sampled in that raster */
-        typedef std::unordered_map<uint32_t, std::vector<uint32_t>> raster_points_map_t;
+        typedef std::unordered_map<uint64_t, std::vector<uint32_t>> raster_points_map_t;
 
         /* GroupsFinder thread info used by batch sampling code */
         typedef struct GroupsFinder {
@@ -162,7 +162,7 @@ class GeoIndexedRaster: public RasterObject
             std::vector<point_groups_t>      pointsGroups;
             raster_points_map_t              rasterToPointsMap;
             RasterFileDictionary             threadFileDict;
-            std::unordered_map<uint32_t, uint32_t> localToGlobalFileIds; // cache for mapping thread-local file ids to global ids
+            std::unordered_map<uint64_t, uint64_t> localToGlobalFileIds; // cache for mapping thread-local file ids to global ids
             std::unordered_map<long, OGRFeature*>  featureCache;         // cache cloned features by FID to avoid repeated clones
 
             explicit GroupsFinder (GeoIndexedRaster* _obj, const std::vector<point_info_t>* _points);
